backend/requirements.txt
fastapi
fastapi[standard]
uvicorn[standard]
redis
asyncio
PyYAML